move: move AddGroupCommand to usecase/command
